push_device [nfc]: Move registering push token off of UpdateMachine
This makes a more comfortable home for adding further logic
in this area.
